feat: comprar e listar, alterar e excluir compra

+++ src/controllers/produto.controllers.js
@@ -3,11 +3,28 @@ import Produto from "../model/produto.js"
 
 
 async function admin(req,res){
-    res.render('tela_de_administracao')
+    if(id == null && !email_admin){
+        res.redirect('/usuario/login')
+    }
+    else if (email_admin && id == null){
+        res.render('tela_de_administracao')
+    }
+    else{
+        res.redirect('/usuario/principal')
+    }
 }
 
 async function formularioProduto(req,res){
-    res.render('TelaCadastro')
+    if(id == null && !email_admin){
+        res.redirect('/usuario/login')
+    }
+    else if (email_admin && id == null){
+        res.render('TelaCadastro')
+    }
+    else{
+        res.redirect('/usuario/principal')
+    }
+
 }
 
 async function cadastrarProduto(req, res){
@@ -15,40 +32,82 @@ async function cadastrarProduto(req, res){
     const produto = new Produto(null, novo.nome, novo.categoria, novo.preco, true)
 
     const retorno = await produtoServices.cadastrarProduto(produto)
-    console.log(retorno)
+    //console.log(retorno)
     res.render('resultado', {retorno: retorno})
 
 }
 
 async function listarProdutos(req,res){
-    const retorno = await produtoServices.listarProdutos()
-    res.render('listarProdutos', {retorno: retorno})
+    if(id == null && !email_admin){
+        res.redirect('/usuario/login')
+    }
+    else if (email_admin && id == null){
+        const retorno = await produtoServices.listarProdutos()
+        res.render('listarProdutos', {retorno: retorno})
+    }
+    else{
+        res.redirect('/usuario/principal')
+    }
 }
 
 async function listarProdutosExcluir(req,res){
-    const retorno = await produtoServices.listarProdutos()
-    res.render('TelaExclusao', {retorno: retorno})
+    if(id == null && !email_admin){
+        res.redirect('/usuario/login')
+    }
+    else if (email_admin && id == null){
+        const retorno = await produtoServices.listarProdutos()
+        res.render('TelaExclusao', {retorno: retorno})
+    }
+    else{
+        res.redirect('/usuario/principal')
+    }
 }
 
 async function listarProdutosAlterar(req,res){
-    const retorno = await produtoServices.listarProdutos()
-    res.render('AlterarProduto', {retorno: retorno})
+    if(id == null && !email_admin){
+        res.redirect('/usuario/login')
+    }
+    else if (email_admin && id == null){
+        const retorno = await produtoServices.listarProdutos()
+        res.render('AlterarProduto', {retorno: retorno})
+    }
+    else{
+        res.redirect('/usuario/principal')
+    }
+
 }
 
 async function excluirProduto(req,res){
-    const id = req.params.id
-    await produtoServices.excluirProduto(id)
-    res.redirect('/produto')
+    if(id == null && !email_admin){
+        res.redirect('/usuario/login')
+    }
+    else if (email_admin && id == null){
+        const id = req.params.id
+        await produtoServices.excluirProduto(id)
+        res.redirect('/produto')
+    }
+    else{
+        res.redirect('/usuario/principal')
+    }
+
 }
 
 async function procurarProdutoeditar(req,res){
-    const id = req.params.id
-    const retorno = await produtoServices.procurarProdutoeditar(id)
-    if (retorno != null){
-        res.render('editarProduto', {retorno: retorno})
+    if(id == null && !email_admin){
+        res.redirect('/usuario/login')
+    }
+    else if (email_admin && id == null){
+        const id = req.params.id
+        const retorno = await produtoServices.procurarProdutoeditar(id)
+        if (retorno != null){
+            res.render('editarProduto', {retorno: retorno})
+        }
+        else{
+            res.redirect('/produto/listar-produto')
+        }
     }
     else{
-        res.redirect('/produto/listar-produto')
+        res.redirect('/usuario/principal')
     }
 }
